Add the disabled k/wasm target to the rest of necessary modules
