+++ compose/animation/animation-graphics/build.gradle
@@ -67,6 +67,7 @@ if(AndroidXComposePlugin.isMultiplatformEnabled(project)) {
         desktop()
         darwin()
         js()
+        wasm()
     }
 
     kotlin {
@@ -94,6 +95,24 @@ if(AndroidXComposePlugin.isMultiplatformEnabled(project)) {
 
             desktopMain.dependencies {
                 implementation(libs.kotlinStdlib)
+            }
+
+            jsWasmMain.dependsOn(commonMain)
+
+            jsMain {
+                dependsOn(jsWasmMain)
+                dependencies {
+                    implementation(kotlin("stdlib-js"))
+                }
+            }
+
+            if (project.kotlinWasmEnabled.toBoolean()) {
+                wasmMain {
+                    dependsOn(jsWasmMain)
+                    dependencies {
+                        implementation(kotlin("stdlib-wasm"))
+                    }
+                }
             }
 
             // TODO(b/214407011): These dependencies leak into instrumented tests as well. If you
