+++ compose/foundation/foundation/src/jsWasmMain/kotlin/androidx/compose/foundation/lazy/layout/Lazy.js.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 The Android Open Source Project
+ * Copyright 2023 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,8 @@
  * limitations under the License.
  */
 
-package androidx.compose.foundation.text
+package androidx.compose.foundation.lazy.layout
 
-internal actual fun timeNowMillis(): Long = kotlinx.browser.window.performance.now().toLong()
+actual fun getDefaultLazyLayoutKey(index: Int): Any = DefaultLazyKey(index)
 
+private data class DefaultLazyKey(private val index: Int)
