+++ compose/foundation/foundation/src/jsWasmMain/kotlin/androidx/compose/foundation/text/ContextMenu.js.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 The Android Open Source Project
+ * Copyright 2023 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
