+++ compose/foundation/foundation/src/jsWasmMain/kotlin/androidx/compose/foundation/text/KeyMapping.jsMain.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 The Android Open Source Project
+ * Copyright 2023 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,15 +14,9 @@
  * limitations under the License.
  */
 
-package androidx.compose.foundation.text.selection
+package androidx.compose.foundation.text
 
-import androidx.compose.runtime.Stable
-import androidx.compose.ui.graphics.Color
+import androidx.compose.foundation.text.KeyMapping
+import androidx.compose.foundation.text.defaultKeyMapping
 
-private val DefaultSelectionColor = Color(0xFF4286F4)
-
-@Stable
-internal actual val DefaultTextSelectionColors = TextSelectionColors(
-    handleColor = DefaultSelectionColor,
-    backgroundColor = DefaultSelectionColor.copy(alpha = 0.4f)
-)
+internal actual val platformDefaultKeyMapping: KeyMapping = defaultKeyMapping
