+++ compose/foundation/foundation/src/jsWasmMain/kotlin/androidx/compose/foundation/text/TextFieldCursor.js.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 The Android Open Source Project
+ * Copyright 2023 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,13 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package androidx.compose.foundation.text
 
-package androidx.compose.foundation.lazy.layout
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
 
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.State
-import androidx.compose.ui.layout.SubcomposeLayoutState
-
-actual fun getDefaultLazyLayoutKey(index: Int): Any = DefaultLazyKey(index)
-
-private data class DefaultLazyKey(private val index: Int)
+internal actual val DefaultCursorThickness: Dp = 1.dp
