+++ compose/foundation/foundation/src/jsWasmMain/kotlin/androidx/compose/foundation/text/TextFieldKeyInput.js.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 The Android Open Source Project
+ * Copyright 2023 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,7 +16,12 @@
 
 package androidx.compose.foundation.text
 
-import androidx.compose.ui.input.pointer.DummyPointerIcon
-import androidx.compose.ui.input.pointer.PointerIcon
+import androidx.compose.ui.input.key.KeyEvent
+import org.jetbrains.skiko.*
 
-internal actual val textPointerIcon: PointerIcon = DummyPointerIcon
+actual val KeyEvent.isTypedEvent: Boolean
+    get() = nativeKeyEvent.kind == SkikoKeyboardEventKind.DOWN &&
+        nativeKeyEvent.platform?.isPrintable() == true
+
+private fun SkikoPlatformKeyboardEvent.isPrintable(): Boolean =
+    charCode != 0
