+++ compose/foundation/foundation/src/jsWasmMain/kotlin/androidx/compose/foundation/text/TextPointerIcon.js.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 The Android Open Source Project
+ * Copyright 2023 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,10 +14,9 @@
  * limitations under the License.
  */
 
-package androidx.compose.foundation.text.selection
+package androidx.compose.foundation.text
 
-import androidx.compose.ui.input.pointer.PointerEvent
-import androidx.compose.ui.Modifier
+import androidx.compose.ui.input.pointer.DummyPointerIcon
+import androidx.compose.ui.input.pointer.PointerIcon
 
-internal actual fun Modifier.textFieldMagnifier(manager: TextFieldSelectionManager): Modifier =
-    TODO("implement js textFieldMagnifier")
+internal actual val textPointerIcon: PointerIcon = DummyPointerIcon
