+++ compose/foundation/foundation/src/jsWasmMain/kotlin/androidx/compose/foundation/text/UndoManager.js.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 The Android Open Source Project
+ * Copyright 2023 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,8 +16,5 @@
 
 package androidx.compose.foundation.text
 
-import androidx.compose.ui.input.key.KeyEvent
+internal actual fun timeNowMillis(): Long = kotlinx.browser.window.performance.now().toLong()
 
-internal actual fun KeyEvent.cancelsTextSelection(): Boolean = false
-
-internal actual fun showCharacterPalette():Unit = TODO("implement native showCharacterPalette")
