+++ compose/foundation/foundation/src/jsWasmMain/kotlin/androidx/compose/foundation/text/selection/DefaultTextSelectionColors.js.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 The Android Open Source Project
+ * Copyright 2023 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,6 +14,15 @@
  * limitations under the License.
  */
 
-package androidx.compose.foundation.text
+package androidx.compose.foundation.text.selection
 
-internal actual val isInTouchMode = false
+import androidx.compose.runtime.Stable
+import androidx.compose.ui.graphics.Color
+
+private val DefaultSelectionColor = Color(0xFF4286F4)
+
+@Stable
+internal actual val DefaultTextSelectionColors = TextSelectionColors(
+    handleColor = DefaultSelectionColor,
+    backgroundColor = DefaultSelectionColor.copy(alpha = 0.4f)
+)
