+++ compose/foundation/foundation/src/jsWasmMain/kotlin/androidx/compose/foundation/text/selection/SelectionHandles.js.kt
