+++ compose/material3/material3/build.gradle
@@ -82,6 +82,7 @@ if(AndroidXComposePlugin.isMultiplatformEnabled(project)) {
         android()
         desktop()
         darwin()
+        wasm()
         js()
     }
 
@@ -126,8 +127,12 @@ if(AndroidXComposePlugin.isMultiplatformEnabled(project)) {
 
             jsNativeMain.dependsOn(skikoMain)
 
-            jsMain.dependsOn(jsNativeMain)
             nativeMain.dependsOn(jsNativeMain)
+            jsWasmMain.dependsOn(jsNativeMain)
+            jsMain.dependsOn(jsWasmMain)
+            if (project.kotlinWasmEnabled.toBoolean()) {
+                wasmMain.dependsOn(jsWasmMain)
+            }
 
             // TODO(b/214407011): These dependencies leak into instrumented tests as well. If you
             //  need to add Robolectric (which must be kept out of androidAndroidTest), use a top
